refactor(services/user): rename to auth
